% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\docType{data}
\name{available_price_indexes}
\alias{available_price_indexes}
\title{Price indexes available in the \code{realtalk} package}
\format{
The data frame \code{available_price_indexes} shows which datasets are
available from the \verb{realtalk package}. It has six columns:
\describe{
\item{data_source}{Short price index name}
\item{frequency}{time frequency of data available)}
\item{seasonal}{seasonally adjustment status of the data}
\item{min_date}{first date of availability}
\item{max_date}{last date of availability}
\item{package_data_name}{internal data name of the series}
}

Use the \code{package_data_name} value to load a specific series.
}
\usage{
available_price_indexes
}
\description{
Price indexes available in the \code{realtalk} package
}
\examples{
available_price_indexes
}
\keyword{datasets}
