% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_cpi_u.R
\docType{data}
\name{c_cpi_u}
\alias{c_cpi_u}
\alias{c_cpi_u_monthly_nsa}
\alias{c_cpi_u_annual}
\title{Chained Consumer Price Index for Urban Consumers (C-CPI-U) data}
\format{
The data frame \code{c_cpi_u_monthly_nsa}
contains not-seasonally adjusted price index levels of the
Chained CPI-U (C-CPI-U). It has three columns:
\describe{
\item{year}{numeric year}
\item{month}{numeric calendar month (1-12)}
\item{c_cpi_u}{value of the C-CPI-U price index}
}

The data frame \code{c_cpi_u_annual} contains the annual
price index level of the Chained CPI-U (C-CPI-U). It has two columns:
\describe{
\item{year}{numeric year}
\item{c_cpi_u}{value of the C-CPI-U price index}
}
}
\source{
\url{https://www.bls.gov/cpi/}
}
\usage{
c_cpi_u_monthly_nsa

c_cpi_u_annual
}
\description{
Monthly or annual C-CPI-U price indexes
}
\examples{
c_cpi_u_monthly_nsa
}
\seealso{
\code{\link[=cpi_u_rs]{cpi_u_rs_monthly_nsa}} and \code{\link[=cpi_u_rs]{cpi_u_rs_annual}}
for the CPI-U-RS, a CPI index that is consistently defined over time and goes
back to 1978.
}
\keyword{datasets}
