% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_cpi_u_extended.R
\docType{data}
\name{c_cpi_u_extended}
\alias{c_cpi_u_extended}
\alias{c_cpi_u_extended_monthly_nsa}
\alias{c_cpi_u_extended_annual}
\title{Extended Chained Chained Consumer Price Index for Urban Consumers (C-CPI-U)}
\format{
The extended C-CPI-U series uses the 1978-2000 CPI-U-RS
and extends it retroactively by chaining it to the
CPI-U-RS (1978-1999), CPI-U-X1 (1967-1977),
and CPI-U (1966 and before).

The data frame \code{c_cpi_u_extended_monthly_nsa} contains
non-seasonally adjusted price index levels of the extended C-CPI-U.
It has three columns:
\describe{
\item{year}{numeric year}
\item{month}{numeric calendar month (1-12)}
\item{pce}{value of the C-CPI-U extended price index}
}

The data frame \verb{c_cpi_u_extended annual} contains the annual
price index level of the extended C-CPI-U. It has two columns:
\describe{
\item{year}{numeric year}
\item{pce}{value of the C-CPI-U extended price index}
}
}
\source{
The U.S. Census Bureau currently uses a similar combination of
price indexes to extend the CPI-U-RS and CPI-U: \url{https://www.census.gov/topics/income-poverty/income/guidance/current-vs-constant-dollars.html}.
}
\usage{
c_cpi_u_extended_monthly_nsa

c_cpi_u_extended_annual
}
\description{
The monthly or annual C-CPI-U series extended to 1937 through
the present, based on CPI-U-RS, CPI-U, and CPI-U-X1.
}
\examples{
c_cpi_u_extended_annual
}
\keyword{datasets}
