% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{capture_audio_chunk_file}
\alias{capture_audio_chunk_file}
\title{Capture an Audio Chunk to File}
\usage{
capture_audio_chunk_file(file, duration = 1, bitrate = 24000)
}
\arguments{
\item{file}{File to write to}

\item{duration}{Duration in seconds to record (default 1 sec).}

\item{bitrate}{Bitrate in Hz (default 24000)}
}
\description{
Capture an Audio Chunk to File
}
