% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{capture_audio_chunks_to_dir}
\alias{capture_audio_chunks_to_dir}
\title{Capture Audio Chunks to a Directory}
\usage{
capture_audio_chunks_to_dir(dir, duration = 1, bitrate = 24000)
}
\arguments{
\item{dir}{Directory to write the wav files to}

\item{duration}{Duration in seconds to record (default 1 sec).}

\item{bitrate}{Bitrate in Hz (default 24000)}
}
\description{
Capture Audio Chunks to a Directory
}
