% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_u.R
\docType{data}
\name{cpi_u}
\alias{cpi_u}
\alias{cpi_u_monthly_nsa}
\alias{cpi_u_monthly_sa}
\alias{cpi_u_quarterly_nsa}
\alias{cpi_u_quarterly_sa}
\alias{cpi_u_annual}
\title{Consumer Price Index for Urban Consumers (CPI-U) data}
\format{
The data frames \code{cpi_u_monthly_sa} and \code{cpi_u_monthly_nsa}
contain, respectively, seasonally adjusted or not seasonally adjusted monthly
price index levels of the CPI-U. They have three columns:
\describe{
\item{year}{numeric year}
\item{month}{numeric calendar month (1-12)}
\item{cpi_u}{value of the CPI-U price index}
}

#' @format The data frames \code{cpi_u_quarterly_sa} and \code{cpi_u_quarterly_nsa}
contain, respectively, seasonally adjusted or not seasonally adjusted quarterly
price index levels of the CPI-U. They have three columns:
\describe{
\item{year}{numeric year}
\item{quarter}{numeric calendar quarter (1-4)}
\item{cpi_u}{value of the CPI-U price index}
}

The data frame \code{cpi_u_annual} contains the annual
price index level of the CPI-U. It has two columns:
\describe{
\item{year}{numeric year}
\item{cpi_u}{value of the CPI-U price index}
}
}
\source{
\url{https://www.bls.gov/cpi/}
}
\usage{
cpi_u_monthly_nsa

cpi_u_monthly_sa

cpi_u_quarterly_nsa

cpi_u_quarterly_sa

cpi_u_annual
}
\description{
Monthly, quarterly, or annual CPI-U price indexes
}
\examples{
cpi_u_monthly_nsa
}
\seealso{
\code{\link[=cpi_u_rs]{cpi_u_rs_monthly_nsa}} and \code{\link[=cpi_u_rs]{cpi_u_rs_annual}}
for the CPI-U-RS, an index that is more consistent over time than the CPI-U
and incorporates more recent improvements to the CPI-U into the
entire time series. The quarterly \code{cpi_u_quarterly_nsa} and
\code{cpi_u_quarterly_sa} series are the quarterly means of their
respective monthly series.
}
\keyword{datasets}
