% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_u_rs.R
\docType{data}
\name{cpi_u_rs}
\alias{cpi_u_rs}
\alias{cpi_u_rs_monthly_nsa}
\alias{cpi_u_rs_annual}
\title{Consumer Price Index for Urban Consumers Research Series (CPI-U-RS) data}
\format{
The data frame \code{cpi_u_rs_monthly_nsa} contains the
not seasonally adjusted monthly price index levels of the CPI-U-RS
(CPI-U research series). The CPI-U-RS is more consistent over time than
the CPI-U and incorporates more recent improvements to the CPI-U into the
entire time series. It has three columns:
\describe{
\item{year}{numeric year}
\item{month}{numeric calendar month (1-12)}
\item{cpi_u_rs}{value of the CPI-U-RS price index}
}

The data frame \code{cpi_u_rs_annual} contains the annual
price index level of the CPI-U-RS. It has two columns:
\describe{
\item{year}{numeric year}
\item{cpi_u_rs}{value of the CPI-U-RS price index}
}
}
\source{
\url{https://www.bls.gov/cpi/research-series/r-cpi-u-rs-home.htm}
}
\usage{
cpi_u_rs_monthly_nsa

cpi_u_rs_annual
}
\description{
Monthly or annual CPI-U-RS price indexes
}
\examples{
cpi_u_rs_annual
}
\seealso{
\code{\link[=c_cpi_u_extended]{c_cpi_u_extended_annual}} extends the
CPI-U-RS from 1937 to the present using the \code{\link{c_cpi_u}} and other series.
}
\keyword{datasets}
