% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_u_x1.R
\docType{data}
\name{cpi_u_x1}
\alias{cpi_u_x1}
\alias{cpi_u_x1_monthly_nsa}
\alias{cpi_u_x1_annual}
\title{Experimental Consumer Price Index for Urban Consumers X1 (CPI-U-X1) data}
\format{
The data frame \code{cpi_u_x1_monthly_nsa} contains the
not seasonally adjusted monthly price index levels of the CPI-U-X1
from January 1967 through December 1982. It has three columns:
\describe{
\item{year}{numeric year}
\item{month}{numeric calendar month (1-12)}
\item{cpi_u_x1}{value of the CPI-U-X1 price index}
}

The data frame \code{cpi_u_x1_annual} contains the annual average
price index level of the CPI-U-X1. It has two columns:
\describe{
\item{year}{numeric year}
\item{cpi_u_x1}{value of the CPI-U-X1 price index}
}
}
\source{
Emailed spreadsheet from the US Bureau of Labor Statistics
}
\usage{
cpi_u_x1_monthly_nsa

cpi_u_x1_annual
}
\description{
Monthly or annual CPI-U-X1 price indexes
}
\examples{
cpi_u_x1_annual
}
\seealso{
\code{\link[=cpi_u_rs_extended]{cpi_u_rs_extended_monthly_nsa}} and \code{\link[=cpi_u_rs_extended]{cpi_u_extended_rs_annual}}
for an extended version of the CPI-U-RS, using the \code{\link{cpi_u}} and \code{\link{cpi_u_x1}}
which is more consistent over time than the CPI-U and incorporates more recent
improvements to the CPI-U into the entire time series.
}
\keyword{datasets}
