% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventlog.R
\name{EventLog}
\alias{EventLog}
\title{R6 Class Representing a Log of Events in a Stream}
\description{
R6 Class Representing a Log of Events in a Stream

R6 Class Representing a Log of Events in a Stream
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EventLog-new}{\code{EventLog$new()}}
\item \href{#method-EventLog-add}{\code{EventLog$add()}}
\item \href{#method-EventLog-as_tibble}{\code{EventLog$as_tibble()}}
\item \href{#method-EventLog-print}{\code{EventLog$print()}}
\item \href{#method-EventLog-clone}{\code{EventLog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventLog-new"></a>}}
\if{latex}{\out{\hypertarget{method-EventLog-new}{}}}
\subsection{Method \code{new()}}{
Create an EventLog
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLog$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventLog-add"></a>}}
\if{latex}{\out{\hypertarget{method-EventLog-add}{}}}
\subsection{Method \code{add()}}{
Add an Event
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLog$add(event)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{An Event object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventLog-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-EventLog-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
Return the eventlog as a tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLog$as_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventLog-print"></a>}}
\if{latex}{\out{\hypertarget{method-EventLog-print}{}}}
\subsection{Method \code{print()}}{
Print the eventlog
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLog$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventLog-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EventLog-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventLog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
