% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_datetime}
\alias{format_datetime}
\title{Format current datetime}
\usage{
format_datetime()
}
\value{
A character vector containing the formatted current datetime
}
\description{
Format current datetime
}
