% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_price_index}
\alias{get_price_index}
\title{Get price index series as a data frame}
\usage{
get_price_index(index_name, frequency, seasonal = NA)
}
\arguments{
\item{index_name}{Name of the price index, as specified by index name in \code{\link{available_price_indexes}}}

\item{frequency}{"monthly" or "annual"}

\item{seasonal}{"NSA" or "SA" for monthly data; NA (default) for annual}
}
\value{
A tibble with the requested price index
}
\description{
Get price index series as a data frame
}
\examples{
get_price_index("CPI-U-RS", "monthly", "NSA")
}
