% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{openai_api_key}
\alias{openai_api_key}
\title{Retrieve OpenAI API key from the OPENAI_API_KEY environmental variable}
\usage{
openai_api_key(verbose = TRUE)
}
\arguments{
\item{verbose}{Print messages? Defaults to TRUE. Errors will always be
printed.}
}
\value{
A character string containing the API key
}
\description{
Retrieve OpenAI API key from the OPENAI_API_KEY environmental variable
}
