% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pce.R
\docType{data}
\name{pce}
\alias{pce}
\alias{pce_monthly_sa}
\alias{pce_annual}
\title{Personal Consumption Expenditures (PCE) price index}
\format{
The data frame \code{pce_monthly_sa} contains seasonally adjusted
price index levels of the PCE. It has three columns:
\describe{
\item{year}{numeric year}
\item{month}{numeric calendar month (1-12)}
\item{pce}{value of the PCE price index}
}

The data frame \code{pce_annual} contains the annual
price index level of the PCE. It has two columns:
\describe{
\item{year}{numeric year}
\item{pce}{value of the PCE price index}
}
}
\source{
\url{https://www.bea.gov/data/personal-consumption-expenditures-price-index}
}
\usage{
pce_monthly_sa

pce_annual
}
\description{
Monthly or annual PCE price indexes
}
\examples{
pce_annual
}
\keyword{datasets}
