% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtalk-package.R
\name{realtalk}
\alias{realtalk}
\alias{realtalk-package}
\title{realtalk: Interface to the OpenAI Realtime API}
\description{
Provides tools for interacting with the OpenAI Realtime API.
Create streaming sessions with real-time text and audio input/output,
handle events, and manage communication with the API through a WebSocket
connection. The package supports bidirectional streaming of text and audio
content in real-time applications.
}
\author{
David Wilkins \href{mailto:david@wilkox.org}{david@wilkox.org}
}
