% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{start_stream}
\alias{start_stream}
\title{Start a Streaming Connection to OpenAI Realtime API via WebSocket}
\usage{
start_stream(
  api_key = lemur::openai_api_key(),
  model = "gpt-4o-realtime-preview-2024-12-17",
  voice = "verse"
)
}
\arguments{
\item{api_key}{Your long-term OpenAI API key. Defaults to
\code{Sys.getenv("OPENAI_API_KEY")}.}

\item{model}{A character string specifying the model. Defaults to
\code{lemur::openai_api_key()}.}

\item{voice}{A character string specifying the voice to use (default is "verse").}
}
\description{
Connects to the realtime endpoint using a WebSocket with the given model and
API key.
}
