% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_u_x1.R
\docType{data}
\name{cpi_u_x1}
\alias{cpi_u_x1}
\alias{cpi_u_x1_monthly_nsa}
\alias{cpi_u_x1_annual}
\title{Experimental Consumer Price Index for Urban Consumers X1 (CPI-U-X1) data}
\format{
The data frame \code{cpi_u_x1_monthly_nsa} contains the
not seasonally adjusted monthly price index levels of the CPI-U-X1
from January 1967 through December 1982. It has three columns:
\describe{
\item{year}{numeric year}
\item{month}{numeric calendar month (1-12)}
\item{cpi_u_x1}{value of the CPI-U-X1 price index}
}

The data frame \code{cpi_u_x1_annual} contains the annual average
price index level of the CPI-U-X1. It has two columns:
\describe{
\item{year}{numeric year}
\item{cpi_u_x1}{value of the CPI-U-X1 price index}
}
}
\source{
Emailed spreadsheet from the US Bureau of Labor Statistics
}
\usage{
cpi_u_x1_monthly_nsa

cpi_u_x1_annual
}
\description{
Monthly or annual CPI-U-X1 price indexes
}
\examples{
cpi_u_x1_annual
}
\seealso{
\code{\link[=c_cpi_u_extended]{c_cpi_u_extended_annual}} extends the
CPI-U-X1 from 1937 to the present using the \code{\link{c_cpi_u}} and other series.
}
\keyword{datasets}
