% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{Event}
\alias{Event}
\title{R6 Class Representing a Stream Event}
\description{
R6 Class Representing a Stream Event

R6 Class Representing a Stream Event
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{The raw event data}

\item{\code{created_at}}{Timestamp for when the Event was created}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Event-new}{\code{Event$new()}}
\item \href{#method-Event-as_tibble}{\code{Event$as_tibble()}}
\item \href{#method-Event-print}{\code{Event$print()}}
\item \href{#method-Event-clone}{\code{Event$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Event-new"></a>}}
\if{latex}{\out{\hypertarget{method-Event-new}{}}}
\subsection{Method \code{new()}}{
Create an Event
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Event$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The event data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Event-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-Event-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
Return the event as a tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Event$as_tibble()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Event-print"></a>}}
\if{latex}{\out{\hypertarget{method-Event-print}{}}}
\subsection{Method \code{print()}}{
Print the event
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Event$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Event-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Event-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Event$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
