% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{play_audio_chunk}
\alias{play_audio_chunk}
\title{Decode Base64 Audio and Play It}
\usage{
play_audio_chunk(audio_b64)
}
\arguments{
\item{audio_b64}{A base64-encoded string representing audio.}
}
\description{
Decode Base64 Audio and Play It
}
