% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtalk-package.R
\docType{package}
\name{realtalk-package}
\alias{realtalk}
\alias{realtalk-package}
\title{realtalk: Price index data for the US economy}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Makes it easy to use US price index data like the CPI.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://economic.github.io/realtalk/}
}

}
\author{
\strong{Maintainer}: Ben Zipperer \email{bzipperer@epi.org}

Other contributors:
\itemize{
  \item Economic Policy Institute [copyright holder, funder]
}

}
\keyword{internal}
